+++ package.json
@@ -12,9 +12,11 @@
     "@testing-library/user-event": "^12.6.2",
     "@types/jest": "^26.0.20",
     "@types/node": "^12.19.15",
+    "@types/pubsub-js": "^1.8.2",
     "@types/react": "^16.14.2",
     "@types/react-dom": "^16.9.10",
     "@types/react-router-dom": "^5.1.7",
+    "@types/websocket": "^1.0.1",
     "graphql": "^15.4.0",
     "i18next": "^19.8.4",
     "keycloak-js": "^12.0.2",
