--- src/qgl/DiagramService.js
@@ -1,7 +1,0 @@
-import * as data from "../mock/RouteNodeDiagramObjectsExample.json";
-
-function DiagramService() {
-  return data;
-}
-
-export default DiagramService;
